Rename main.h to fast_lio_sam_qn.h
