+++ fast_lio_sam_qn/src/main.cpp
@@ -1,4 +1,4 @@
-#include "main.h"
+#include "fast_lio_sam_qn.h"
 
 int main(int argc, char **argv)
 {
